target/fa626: Added FA626 driver
The details of this driver were derrived from a decompiled build of
openocd contained within a copy of the ITE Castor3 SDK for the
IT9070-series processors which contain a Faraday Technology FA626:

https://github.com/jhol/20171003_Doorbell_SDK_v1.2.3.1_758862/blob/master/ITE_Castor3_SDK/tool/bin/openocd.exe

The decompilation work was done with Ghidra.

By comparing the decompilation of both the FA626 and FA526 drivers,
alongside the public OpenOCD FA526 driver code it was possible to
determine that the two processor cores closely resemble each other.

The only differences found were that the FA626 has a longer execution
pipeline, and it has support for ARM Thumb which the FA526 lacks.